+++ src/main/kotlin/racing/domain/car/RacingCars.kt
@@ -1,4 +1,6 @@
-package racing.domain
+package racing.domain.car
+
+import racing.domain.movement.MoveStrategy
 
 class RacingCars private constructor(private val racingCars: List<RacingCar>) {
     companion object {
