+++ src/main/kotlin/racing/domain/movement/MoveStrategy.kt
@@ -1,4 +1,4 @@
-package racing.domain
+package racing.domain.movement
 
 interface MoveStrategy {
     val decisionNumber: Int
